C1: close the clippy --all-targets gate — 339 hidden lint sites cleared, the gap recorded since i349 is shut
plan-rust-craft C1 (docs/PLAN_RUST_CRAFT_AUDIT.md). Measured before: clippy
--workspace --all-targets was 1 hard error + 339 warning sites across 147
files, none of which the gate could see (it ran --quiet and grepped output).
The gate step and the versioned pre-commit hook now run
`--all-targets --locked -- -D warnings` — the compiler's exit code, not a grep
— and the AGENTS.md §3 known-gap paragraph is amended to its closed form.

Cleared in the same commit: the F1 dead probe (i286_norm_proposal_census; its
`* 0` erasing_op blocks a -D warnings gate — corrected at step 0, plain rustc
accepted it, so the census did run and the evidence doc is annotated rather
than retracted), the F3 lint-inheritance hole (mindstrata-development now
inherits the workspace lint set, 0 new warnings as simulated), and every
cfg(test)/example warning. One machine-applied --fix needed hand-repair:
i387_utility_decomposition kept a `*winner` deref after the fix hoisted the
value into a local (E0614) — the reason step 6 is the point of the iteration.

Structural only, no production behaviour changed. Proof: gate GREEN with both
goldens byte-identical, integration 314/0/1 unchanged, sim 310/310.

+++ crates/mindstrata-benches/examples/i280_violations_sweep.rs
@@ -6,7 +6,7 @@
 use mindstrata_sim::sim::{SimConfig, Simulation};
 
 fn violations(sim: &Simulation) -> usize {
-    sim.recent_events(sim.event_count() as usize)
+    sim.recent_events(sim.event_count())
         .iter()
         .filter(|e| matches!(e, mindstrata_core::event::SimEvent::NormViolated { .. }))
         .count()
