+++ crates/mindstrata-benches/examples/i288_transgression_verdict.rs
@@ -30,7 +30,7 @@ fn census(seed: u64, ticks: u64) {
     let mut sim = Simulation::new(config(seed, ticks));
     sim.populate();
     sim.run(ticks);
-    let events = sim.recent_events(sim.event_count() as usize);
+    let events = sim.recent_events(sim.event_count());
     let norm_violated = events
         .iter()
         .filter(|e| matches!(e, SimEvent::NormViolated { .. }))
